+++ package.json
@@ -12,12 +12,12 @@
   },
   "devDependencies": {
     "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^14.1.0",
+    "eslint-config-airbnb": "^15.0.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "4.0.0",
-    "eslint-plugin-react": "6.9.0",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
     "husky": "^0.13.3",
-    "react-scripts": "0.9.5"
+    "react-scripts": "^1.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
