+++ src/App.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import styled from 'styled-components';
 import 'normalize.css';
-import ProductList from './ProductList';
-import ProductDetails from './ProductDetails';
+
+import ProductList from './Products/List';
+import ProductDetails from './Products/Show';
 import Sidebar from './Sidebar';
 import './css/fonts.css';
 
@@ -11,8 +12,11 @@ const Wrapper = styled.div`
   min-height: 100vh;
   width: 100%;
   margin: 0px;
-  display: flex;
-  flex-flow: row nowrap;
+  display: block;
+  @media screen and (min-width: 768px) {
+    display: flex;
+    flex-flow: row nowrap;
+  }
 `;
 
 export default () => (
