--- src/ProductDetails/Viewer/Preview/MorePhotos/index.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const Wrapper = styled.p`
-  font-family: "AvenirNextReg";
-  font-size: 14px;
-  color: #c0c0c0;
-  margin-right: 54px;
-`;
-
-export default () => (
-  <Wrapper>
-      see <br /> more <br /> photos
-    </Wrapper>
-  );
