+++ src/Products/List/Filters/Label.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export default styled.span`
+  padding-right: 20px;
+  font: 24px/1.375 "AvenirNext";
+  text-transform: uppercase;
+  color: #4d42f8;
+  @media screen and (min-width: 768px) {
+    padding-left: 35px;
+    padding-right: 20px;
+  }
+`;
