+++ src/Sidebar/index.jsx
@@ -1,22 +1,40 @@
 import React from 'react';
 import styled from 'styled-components';
+
 import Logo from './Logo';
 import Search from './Search';
-import Categories from './Сategories';
+import Navigation from './Navigation';
+import Menu from './Menu';
 
 const Wrapper = styled.aside`
-  background-color: #0e0e0e;
-  flex: 1 0 414px;
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  left: 0;
+  height: 65px;
+  width: 100%;
   display: flex;
-  flex-flow: row wrap;
-  justify-content: center;
-  align-content: flex-start;
+  justify-content: space-between;
+  background-color: #0e0e0e;
+  @media screen and (min-width: 768px) {
+    position: static;
+    height: initial;
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+    align-content: flex-start;
+    flex: 1 0 327px;
+  }
+  @media screen and (min-width: 1024px) {
+    flex: 1 0 414px;
+  }
 `;
 
 export default () => (
   <Wrapper>
     <Logo />
+    <Menu />
     <Search />
-    <Categories />
+    <Navigation />
   </Wrapper>
   );
