+++ src/components/Image.jsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export default styled.img`
+  width: 100%;
+`;
