+++ src/components/Label.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export default styled.span`
+  font: normal normal 14px/1.35 "AvenirNext";
+  text-transform: uppercase;
+  color: #ffffff;
+  background-color: #ff5c5c;  
+`;
